requirements.txt
selenium==4.19.1
pytest==8.1.1
pytest-html==4.1.1
pytesseract==0.3.10
Pillow==10.2.0
webdriver-manager==4.0.1
